+++ CHANGELOG.md
@@ -14,6 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [api] Fix incorrect key property when creating JWT in OIDC identity provider
 - [core] Do not assume preseed.base is a string, handle objects too
 
+### Removed
+
+- [api] We no longer have a dedicated `/preseed` endpoint, use the `/setup` endpoint instead
+- [api] We no longer have a dedicated `/validate/preseed` endpoint, use the `/validate/settings` endpoint instead
+- [core] We no longer have a dedicated `/preseed` endpoint, use the `/setup` endpoint instead
+
 ## [0.5.1] - 2024-11-12
 
 ### Added